Make ajax method compatible with nextRecordsUrl
The ajax call no longer prepends /services/data to paths that already
start with it.  This allows the ajax method to be used with the URL
contained in the nextRecordsUrl attribute of the result object passed to
the query method's success callback.

Here's an example:

forcetkClient.query('SELECT Id, Name FROM Account', handleResult, function() {});

function handleResult(result) {
	// do something with the result
	if (! result.done) {
		forcetkClient.ajax(result.nextRecordsUrl, handleResult, function() {});
	}
}